Replace long with int
Python 3.x no longer has a separate long integer type (integers handle this
under the covers). This fixes exceptions such as...

  Traceback (most recent call last):
    File "/usr/local/lib/python3.5/dist-packages/twisted/trial/runner.py", line 823, in loadByName
      return self.suiteFactory([self.findByName(name, recurse=recurse)])
    File "/usr/local/lib/python3.5/dist-packages/twisted/trial/runner.py", line 702, in findByName
      __import__(name)
    File "/home/atagar/Desktop/tor/bridgedb/bridgedb/test/test_Storage.py", line 13, in <module>
      import bridgedb.Storage as Storage
    File "/home/atagar/Desktop/tor/bridgedb/bridgedb/Storage.py", line 16, in <module>
      from bridgedb.Stability import BridgeHistory
    File "/home/atagar/Desktop/tor/bridgedb/bridgedb/Stability.py", line 36, in <module>
      discountIntervalMillis = long(60*60*12*1000)
  builtins.NameError: name 'long' is not defined

This didn't change the test results much...

  before: FAILED (skips=1, failures=7, errors=49, successes=255)
  after:  FAILED (skips=1, failures=7, errors=48, successes=256)

+++ bridgedb/Bridges.py
@@ -374,7 +374,7 @@ class FixedBridgeSplitter(object):
     def insert(self, bridge):
         # Grab the first 4 bytes
         digest = self.hmac(bridge.identity)
-        pos = long( digest[:8], 16 )
+        pos = int( digest[:8], 16 )
         which = pos % len(self.rings)
         self.rings[which].insert(bridge)
