+++ bridgedb/test/test_parse_addr.py
@@ -735,7 +735,7 @@ class PortListTest(unittest.TestCase):
         """Test ``__getitem__`` with a string."""
         ports = (443, 9001, 9030)
         portList = addr.PortList(*ports)
-        self.assertEqual(portList.__getitem__(long(0)), 9001)
+        self.assertEqual(portList.__getitem__(0), 9001)
 
     def test_mixedArgs(self):
         """Create a :class:`addr.PortList` with mixed type parameters."""
